improved speed of deb-install.sh, changed ansible venv

+++ zsh/.zsh/functions.zsh
@@ -3,14 +3,33 @@
 
 #VENV Python COMMANDS
 function venv-on(){
-  source ./venv/bin/activate
+  if [ -d "./venv" ]; then
+    source ./venv/bin/activate
+    #argcomplete(Ansible)
+    if [ -f "./ansible.cfg" ]; then
+      echo "󱂚 ansible"
+      autoload -Uz compinit
+      compinit
+      eval "$(register-python-argcomplete ansible)"
+    fi
+  else
+    echo "Virtual environment not found."
+  fi
 }
 
 function venv-create(){
-  python3 -m venv venv
-  source ./venv/bin/activate
-  pip install --upgrade pip
-  pip install pynvim
+  local dir=$(basename $(pwd))
+  echo -n "Create and activate virtual environment at $dir? [Y/n] "
+  read answer
+
+  if [[ $answer =~ ^[Yy]$ || $answer == "" ]]; then
+    python3 -m venv venv
+    source ./venv/bin/activate
+    pip install --upgrade pip
+    pip install pynvim
+  else
+    echo "Virtual environment creation aborted."
+  fi
 }
